pyproject.toml
[project]
name = "phylo-gnn"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "biopython>=1.86",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "rapidfuzz>=3.14.3",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "tensorboard>=2.20.0",
    "termine>=0.1.5",
    "torch>=2.9.1",
    "torch-geometric>=2.7.0",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
